Improve .pth install heuristics
Since we can't tell where exactly the package will get installed when
there are Debian installations where the system python is using
dist-packages and pip uses site-packages, we try to install in
any purelib directory that exists.
To offset installing in directories that may not contain the package
itself afterwads (and that can lead to ImportError and an ugly console
message when the site module executes the .pth file) we also check
in the .pth file if the package exists.
Note that this won't work for frozen packages though.

Bump version to v1.1.10.